backend/requirements.txt
uvicorn
fastapi
pydantic
python-multipart
asyncpg
pypdf
python-docx
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.3.0
langchain-postgres>=0.0.12
langchain-text-splitters>=0.3.9
langsmith>=0.1.0
psycopg[binary]>=3.1.0
psycopg[pool]>=3.1.0
